feat(checkout): implement transaction processing and UI improvements
- Add price field to ProductTransaction model and specify table name
- Update product price and stock fields to unsigned types in migration
- Remove cart pagination and product display from checkout view
- Implement transaction processing with product stock updates
- Improve checkout detail page layout and messaging
- Add transaction confirmation options based on payment type

+++ database/migrations/2025_05_15_061856_create_products_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('image');
-            $table->integer('price');
-            $table->integer('stock');
+            $table->unsignedBigInteger('price');
+            $table->unsignedInteger('stock');
             $table->foreignId('brand_id')
                 ->constrained('brands')
                 ->cascadeOnDelete();
